moved Curve stuff to seperate package as it grew too big with all those interfaces; added CurveCombinator as factory for a Curve combined from two other curves
