+++ ImagingKit_Filter/src/main/java/hageldave/imagingkit/filter/util/MiscUtils.java
@@ -11,11 +11,11 @@ public class MiscUtils {
 		return false;
 	}
 	
-	public double clamp(double lowerBound, double value, double upperBound){
+	public static double clamp(double lowerBound, double value, double upperBound){
 		return Math.max(lowerBound, Math.min(value, upperBound));
 	}
 
-	public int clamp(int lowerBound, int value, int upperBound){
+	public static int clamp(int lowerBound, int value, int upperBound){
 		return Math.max(lowerBound, Math.min(value, upperBound));
 	}
